DRILL-5431: SSL Support (C++) - Refactoring of C++ client.
    Move classes out of drillclient to their own files
    Fix build on MacOS to suppress warnings from boost code
    Refactoring of user properties to use a map
