+++ contrib/native/client/src/include/drill/drillClient.hpp
@@ -35,9 +35,12 @@ namespace exec{
 namespace Drill{
 
 //struct UserServerEndPoint;
+class  DrillClientConfig;
+class  DrillClientError;
 class  DrillClientImplBase;
 class  DrillClientImpl;
 class  DrillClientQueryResult;
+class  DrillUserProperties;
 class  FieldMetadata;
 class  PreparedStatement;
 class  RecordBatch;
@@ -49,149 +52,11 @@ enum QueryType{
     PHYSICAL = 3
 };
 
-class DECLSPEC_DRILL_CLIENT DrillClientError{
-    public:
-        static const uint32_t CONN_ERROR_START = 100;
-        static const uint32_t QRY_ERROR_START =  200;
-
-        DrillClientError(uint32_t s, uint32_t e, char* m){status=s; errnum=e; msg=m;};
-        DrillClientError(uint32_t s, uint32_t e, std::string m){status=s; errnum=e; msg=m;};
-
-        static DrillClientError*  getErrorObject(const exec::shared::DrillPBError& e);
-
-        // To get the error number we add a error range start number to
-        // the status code returned (either status_t or connectionStatus_t)
-        uint32_t status; // could be either status_t or connectionStatus_t
-        uint32_t errnum;
-        std::string msg;
-};
-
 // Only one instance of this class exists. A static member of DrillClientImpl;
 class DECLSPEC_DRILL_CLIENT DrillClientInitializer{
     public:
         DrillClientInitializer();
         ~DrillClientInitializer();
-};
-
-// Only one instance of this class exists. A static member of DrillClientImpl;
-class DECLSPEC_DRILL_CLIENT DrillClientConfig{
-    public:
-        DrillClientConfig();
-        ~DrillClientConfig();
-        static void initLogging(const char* path);
-        static void setLogLevel(logLevel_t l);
-        static void setSaslPluginPath(const char* path);
-        static const char* getSaslPluginPath();
-        static void setBufferLimit(uint64_t l);
-        static uint64_t getBufferLimit();
-        static void setSocketTimeout(int32_t l);
-        static void setHandshakeTimeout(int32_t l);
-        static void setQueryTimeout(int32_t l);
-        static void setHeartbeatFrequency(int32_t l);
-        static int32_t getSocketTimeout();
-        static int32_t getHandshakeTimeout();
-        static int32_t getQueryTimeout();
-        static int32_t getHeartbeatFrequency();
-        static logLevel_t getLogLevel();
-
-        /**
-         * Return the client name sent to the server when connecting
-         *
-         * @return the current client name
-         */
-        static const std::string& getClientName();
-
-        /**
-         * Set the client name to be sent to the server when connecting.
-         *
-         * Only new connections will use the new value. Existing connections
-         * will be left unchanged.
-         *
-         * @param name the name to be send to the server
-         */
-        static void setClientName(const std::string& name);
-
-        /**
-         * Return the application name sent to the server when connecting
-         *
-         * @return the current application name
-         */
-        static const std::string& getApplicationName();
-
-        /**
-         * Set the application name to be sent to the server when connecting.
-         *
-         * Only new connections will use the new value. Existing connections
-         * will be left unchanged.
-         *
-         * @param name the name to be send to the server
-         */
-        static void setApplicationName(const std::string& name);
-
-
-
-    private:
-        // The logging level
-        static logLevel_t s_logLevel;
-        // The total amount of memory to be allocated by an instance of DrillClient.
-        // For future use. Currently, not enforced.
-        static uint64_t s_bufferLimit;
-
-        static const char* s_saslPluginPath;
-
-        /**
-         * DrillClient configures timeout (in seconds) in a fine granularity.
-         * Disabled by setting the value to zero.
-         *
-         * s_socketTimout: (default 0)
-         *      set SO_RCVTIMEO and SO_SNDTIMEO socket options and place a
-         *		timeout on socket receives and sends. It is disabled by default.
-         *
-         * s_handshakeTimeout: (default 5)
-         *      place a timeout on validating handshake. When an endpoint (host:port)
-         *		is reachable but drillbit hangs or running another service. It will
-         *		avoid the client hanging.
-         *
-         * s_queryTimeout: (default 180)
-         *      place a timeout on waiting result of querying.
-         *
-         * s_heartbeatFrequency: (default 30)
-         *      Seconds of idle activity after which a heartbeat is sent to the drillbit
-         */
-        static int32_t s_socketTimeout;
-        static int32_t s_handshakeTimeout;
-        static int32_t s_queryTimeout;
-        static int32_t s_heartbeatFrequency;
-        static boost::mutex s_mutex;
-
-        // The client name (default to DRILL_CONNECTOR_NAME)
-        static std::string s_clientName;
-        // The application name (default to <empty>)
-        static std::string s_applicationName;
-};
-
-
-class DECLSPEC_DRILL_CLIENT DrillUserProperties{
-    public:
-        static const std::map<std::string, uint32_t> USER_PROPERTIES;
-
-        DrillUserProperties(){};
-
-        void setProperty( const std::string& propName, const std::string& propValue){
-            std::pair< std::string, std::string> in = make_pair(propName, propValue);
-            m_properties.push_back(in);
-        }
-
-        size_t size() const { return m_properties.size(); }
-
-        const std::string& keyAt(size_t i) const { return m_properties.at(i).first; }
-
-        const std::string& valueAt(size_t i) const { return m_properties.at(i).second; }
-
-        bool validate(std::string& err);
-
-    private:
-        std::vector< std::pair< std::string, std::string> > m_properties;
 };
 
 /*
@@ -1248,7 +1113,7 @@ class DECLSPEC_DRILL_CLIENT DrillClient{
          */
         DEPRECATED connectionStatus_t connect(const char* connectStr, const char* defaultSchema=NULL);
 
-        /*
+        /*  
          * Connect the client to a Drillbit using connection string and a set of user properties.
          * The connection string format can be found in comments of
          * [DRILL-780](https://issues.apache.org/jira/browse/DRILL-780)
